rename contracts -> addresses
