+++ frontend/store/auctions.ts
@@ -10,7 +10,7 @@ import {
 import { checkAllExchangeRates } from 'auctions-core/src/uniswap';
 import { enrichAuctionWithTransactionFees } from 'auctions-core/src/fees';
 import { checkAllCalcParameters } from 'auctions-core/src/params';
-import { checkAllSupportedCollaterals } from 'auctions-core/src/contracts';
+import { checkAllSupportedCollaterals } from 'auctions-core/src/addresses';
 import getWallet from '~/lib/wallet';
 import notifier from '~/lib/notifier';
